fix(ci): restore static analysis gates
