+++ native/rust/crates/ripdpi-proxy-config/src/tests.rs
@@ -10,2116 +10,12 @@ use ripdpi_packets::{MH_DMIX, MH_HMIX, MH_HOSTEXTRASPACE, MH_HOSTTAB, MH_METHODE
 
 use super::*;
 
-fn minimal_ui() -> ProxyUiConfig {
-    let mut config = ProxyUiConfig::default();
-    config.protocols.desync_udp = true;
-    config.chains.tcp_steps = vec![tcp_step("disorder", "host+1")];
-    config.fake_packets.fake_sni = "www.wikipedia.org".to_string();
-    config.host_autolearn.max_hosts = HOST_AUTOLEARN_DEFAULT_MAX_HOSTS;
-    config
-}
-
-fn with_destination_routing_digest(mut json: serde_json::Value) -> serde_json::Value {
-    let section = json.get("destinationRouting").cloned().expect("destinationRouting section");
-    let config: ProxyUiDestinationRoutingConfig = serde_json::from_value(section).expect("destination routing config");
-    let digest = crate::convert::compute_canonical_digest(&config.rules);
-    json["destinationRouting"]["canonicalDigest"] = serde_json::Value::String(digest);
-    json
-}
-
-fn destination_policy_with_digest(rules: Vec<ProxyUiDestinationRoutingRule>) -> ProxyUiDestinationRoutingConfig {
-    let canonical_digest = crate::convert::compute_canonical_digest(&rules);
-    ProxyUiDestinationRoutingConfig {
-        rules,
-        default_action: ProxyUiDestinationRoutingAction::Tunneled,
-        canonical_digest,
-    }
-}
-
-fn exact_aggregate_boundary_policy() -> ProxyUiDestinationRoutingConfig {
-    let rules = (0..256)
-        .map(|rule_index| ProxyUiDestinationRoutingRule {
-            action: ProxyUiDestinationRoutingAction::Direct,
-            network: ProxyUiDestinationRoutingNetwork::Both,
-            domains: (0..4)
-                .map(|matcher_index| {
-                    let prefix = format!("r{rule_index}m{matcher_index}");
-                    ProxyUiDestinationDomainMatcher {
-                        kind: ProxyUiDestinationDomainMatcherKind::Geosite,
-                        value: format!("{prefix:a<63}"),
-                    }
-                })
-                .collect(),
-            ip_ranges: Vec::new(),
-            destination_ports: Vec::new(),
-        })
-        .collect();
-    destination_policy_with_digest(rules)
-}
-
-fn runtime_config_from_destination_policy(
-    destination_routing: ProxyUiDestinationRoutingConfig,
-) -> Result<RuntimeConfig, ProxyConfigError> {
-    runtime_config_from_ui(ProxyUiConfig { destination_routing, ..ProxyUiConfig::default() })
-}
-
-#[test]
-fn shared_destination_routing_fixture_has_cross_language_digest() {
-    let path =
-        golden_test_support::repo_root().join("core/engine/src/test/resources/fixtures/destination-routing-mixed.json");
-    let json: serde_json::Value = serde_json::from_str(&std::fs::read_to_string(path).expect("read shared fixture"))
-        .expect("parse shared fixture");
-    let config: ProxyUiDestinationRoutingConfig =
-        serde_json::from_value(json["destinationRouting"].clone()).expect("destination routing config");
-
-    assert_eq!(
-        crate::convert::compute_canonical_digest(&config.rules),
-        "e7ed9f9ec8688b89eea6f22a7ae6e93e7f441a903b9f9de96d387700c122bee7"
-    );
-}
-
-#[test]
-fn ui_config_maps_geo_database_paths() {
-    let mut ui = minimal_ui();
-    ui.geoip_db_path = Some(" /data/user/0/app/files/geo/geoip.db ".to_string());
-    ui.geosite_db_path = Some("/data/user/0/app/files/geo/geosite.db".to_string());
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert_eq!(config.process.geoip_db_path.as_deref(), Some("/data/user/0/app/files/geo/geoip.db"));
-    assert_eq!(config.process.geosite_db_path.as_deref(), Some("/data/user/0/app/files/geo/geosite.db"));
-}
-
-#[test]
-fn ui_listen_mixed_flag_enables_mixed_network_mode() {
-    let mut ui = minimal_ui();
-    ui.listen.mixed = true;
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert!(config.network.mixed, "listen.mixed must propagate to network.mixed");
-
-    let off = runtime_config_from_ui(minimal_ui()).expect("runtime config");
-    assert!(!off.network.mixed, "mixed must default off when listen.mixed is unset");
-}
-
-fn tcp_step(kind: &str, marker: &str) -> ProxyUiTcpChainStep {
-    ProxyUiTcpChainStep {
-        kind: kind.to_string(),
-        marker: marker.to_string(),
-        midhost_marker: String::new(),
-        fake_host_template: String::new(),
-        fake_order: String::new(),
-        fake_seq_mode: String::new(),
-        tcp_flags_set: String::new(),
-        tcp_flags_unset: String::new(),
-        tcp_flags_orig_set: String::new(),
-        tcp_flags_orig_unset: String::new(),
-        overlap_size: 0,
-        fake_mode: String::new(),
-        fragment_count: 0,
-        min_fragment_size: 0,
-        max_fragment_size: 0,
-        inter_segment_delay_ms: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-        random_fake_host: false,
-    }
-}
-
-fn seqovl_step(marker: &str, overlap_size: i32, fake_mode: &str) -> ProxyUiTcpChainStep {
-    ProxyUiTcpChainStep {
-        kind: "seqovl".to_string(),
-        marker: marker.to_string(),
-        midhost_marker: String::new(),
-        fake_host_template: String::new(),
-        fake_order: String::new(),
-        fake_seq_mode: String::new(),
-        tcp_flags_set: String::new(),
-        tcp_flags_unset: String::new(),
-        tcp_flags_orig_set: String::new(),
-        tcp_flags_orig_unset: String::new(),
-        overlap_size,
-        fake_mode: fake_mode.to_string(),
-        fragment_count: 0,
-        min_fragment_size: 0,
-        max_fragment_size: 0,
-        inter_segment_delay_ms: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-        random_fake_host: false,
-    }
-}
-
-fn udp_step(kind: &str, count: i32) -> ProxyUiUdpChainStep {
-    ProxyUiUdpChainStep {
-        kind: kind.to_string(),
-        count,
-        split_bytes: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-    }
-}
-
-fn activation_filter_with_tcp_state() -> ProxyUiActivationFilter {
-    ProxyUiActivationFilter {
-        round: None,
-        payload_size: None,
-        stream_bytes: None,
-        tcp_has_timestamp: Some(true),
-        tcp_has_ech: Some(false),
-        tcp_window_below: Some(4096),
-        tcp_mss_below: Some(1400),
-    }
-}
-
-fn ui_payload(config: ProxyUiConfig) -> ProxyConfigPayload {
-    ProxyConfigPayload::Ui {
-        strategy_preset: None,
-        config,
-        runtime_context: None,
-        log_context: None,
-        session_overrides: None,
-        schema_version: 2,
-    }
-}
-
-#[test]
-fn ui_payload_parses_hostfake_and_quic_profile() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![ProxyUiTcpChainStep {
-        kind: "hostfake".to_string(),
-        marker: "endhost+8".to_string(),
-        midhost_marker: "midsld".to_string(),
-        fake_host_template: "googlevideo.com".to_string(),
-        fake_order: String::new(),
-        fake_seq_mode: String::new(),
-        tcp_flags_set: String::new(),
-        tcp_flags_unset: String::new(),
-        tcp_flags_orig_set: String::new(),
-        tcp_flags_orig_unset: String::new(),
-        overlap_size: 0,
-        fake_mode: String::new(),
-        fragment_count: 0,
-        min_fragment_size: 0,
-        max_fragment_size: 0,
-        inter_segment_delay_ms: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-        random_fake_host: false,
-    }];
-    ui.chains.udp_steps.push(udp_step("fake_burst", 3));
-    ui.quic.fake_profile = "realistic_initial".to_string();
-    ui.quic.fake_host = "Example.COM.".to_string();
-    ui.fake_packets.tls_fingerprint_profile = "chrome_stable".to_string();
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert_eq!(config.groups[0].actions.http_fake_profile, HttpFakeProfile::CompatDefault);
-    assert_eq!(config.groups[0].actions.tls_fake_profile, TlsFakeProfile::CompatDefault);
-    assert_eq!(config.groups[0].actions.udp_fake_profile, UdpFakeProfile::CompatDefault);
-    assert_eq!(config.groups[0].actions.quic_fake_profile, QuicFakeProfile::RealisticInitial);
-    assert_eq!(config.groups[0].actions.quic_fake_host.as_deref(), Some("example.com"));
-    assert_eq!(config.groups[0].actions.tcp_chain.len(), 2);
-    assert_eq!(config.groups[0].actions.tcp_chain[0].kind(), TcpChainStepKind::TlsRec);
-    assert_eq!(config.groups[0].actions.tcp_chain[0].offset().base, OffsetBase::ExtLen);
-    assert_eq!(config.groups[0].actions.tcp_chain[0].offset().proto, OffsetProto::TlsOnly);
-    assert_eq!(config.groups[0].actions.tcp_chain[1].kind(), TcpChainStepKind::HostFake);
-    assert_eq!(config.groups[0].actions.udp_chain[0].count, 3);
-}
-
-#[test]
-fn ui_payload_without_preset_still_gets_fallback_groups() {
-    let ui = minimal_ui();
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    // Primary group + 3 fallback groups + CONNECT passthrough = at least 5
-    assert!(
-        config.groups.len() >= 5,
-        "strategy_preset=None must still inject ripdpi_default fallback groups, got {} groups",
-        config.groups.len(),
-    );
-    let labels: Vec<&str> = config.groups.iter().map(|g| g.policy.label.as_str()).collect();
-    assert!(labels.contains(&"tlsrec_disorder"), "missing tlsrec_disorder fallback: {labels:?}");
-    assert!(labels.contains(&"disorder_host"), "missing disorder_host fallback: {labels:?}");
-    assert!(labels.contains(&"split_host"), "missing split_host fallback: {labels:?}");
-}
-
-#[test]
-fn ui_payload_preserves_explicit_tlsrec_before_hostfake() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![
-        tcp_step("tlsrec", "extlen"),
-        ProxyUiTcpChainStep {
-            kind: "hostfake".to_string(),
-            marker: "endhost+8".to_string(),
-            midhost_marker: "midsld".to_string(),
-            fake_host_template: "googlevideo.com".to_string(),
-            fake_order: String::new(),
-            fake_seq_mode: String::new(),
-            tcp_flags_set: String::new(),
-            tcp_flags_unset: String::new(),
-            tcp_flags_orig_set: String::new(),
-            tcp_flags_orig_unset: String::new(),
-            overlap_size: 0,
-            fake_mode: String::new(),
-            fragment_count: 0,
-            min_fragment_size: 0,
-            max_fragment_size: 0,
-            inter_segment_delay_ms: 0,
-            activation_filter: None,
-            ipv6_extension_profile: "none".to_string(),
-            random_fake_host: false,
-        },
-    ];
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert_eq!(config.groups[0].actions.tcp_chain.len(), 2);
-    assert_eq!(config.groups[0].actions.tcp_chain[0].kind(), TcpChainStepKind::TlsRec);
-    assert_eq!(config.groups[0].actions.tcp_chain[1].kind(), TcpChainStepKind::HostFake);
-}
-
-#[test]
-fn ui_payload_parses_seqovl_step_and_fields() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("tlsrec", "extlen"), seqovl_step("auto(midsld)", 14, "rand")];
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    let tcp_chain = &config.groups[0].actions.tcp_chain;
-
-    assert_eq!(tcp_chain.len(), 2);
-    assert_eq!(tcp_chain[0].kind(), TcpChainStepKind::TlsRec);
-    assert_eq!(tcp_chain[1].kind(), TcpChainStepKind::SeqOverlap);
-    assert_eq!(tcp_chain[1].offset(), OffsetExpr::adaptive(OffsetBase::AutoMidSld));
-    let payload = tcp_chain[1].seq_overlap_payload().expect("seqovl payload");
-    assert_eq!(payload.overlap_size, 14);
-    assert_eq!(payload.fake_mode, ripdpi_config::SeqOverlapFakeMode::Rand);
-}
-
-#[test]
-fn ui_payload_rejects_duplicate_seqovl_step() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![seqovl_step("host+1", 12, "profile"), seqovl_step("midsld", 16, "rand")];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("duplicate seqovl");
-
-    assert!(err.to_string().contains("seqovl must appear at most once per tcp chain"));
-}
-
-#[test]
-fn ui_payload_rejects_non_leading_seqovl_step() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("split", "host+1"), seqovl_step("midsld", 12, "profile")];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("non-leading seqovl");
-
-    assert!(err.to_string().contains("seqovl must be the first tcp send step"));
-}
-
-#[test]
-fn ui_payload_rejects_tcp_state_predicates_on_group_activation_filter() {
-    let mut ui = minimal_ui();
-    ui.chains.group_activation_filter = Some(activation_filter_with_tcp_state());
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("group tcp-state predicates should be rejected");
-
-    assert!(err.to_string().contains("chains.groupActivationFilter must not declare TCP-state predicates"));
-}
-
-#[test]
-fn ui_payload_rejects_tcp_state_predicates_on_udp_steps() {
-    let mut ui = minimal_ui();
-    ui.chains.udp_steps = vec![ProxyUiUdpChainStep {
-        activation_filter: Some(activation_filter_with_tcp_state()),
-        ..udp_step("fake_burst", 1)
-    }];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("udp tcp-state predicates should be rejected");
-
-    assert!(err.to_string().contains("chains.udpSteps.activationFilter must not declare TCP-state predicates"));
-}
-
-#[test]
-fn ui_payload_parses_multidisorder_terminal_run() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps =
-        vec![tcp_step("tlsrec", "extlen"), tcp_step("multidisorder", "sniext"), tcp_step("multidisorder", "host")];
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    let tcp_chain = &config.groups[0].actions.tcp_chain;
-
-    assert_eq!(tcp_chain.len(), 3);
-    assert_eq!(tcp_chain[0].kind(), TcpChainStepKind::TlsRec);
-    assert_eq!(tcp_chain[1].kind(), TcpChainStepKind::MultiDisorder);
-    assert_eq!(tcp_chain[2].kind(), TcpChainStepKind::MultiDisorder);
-}
-
-#[test]
-fn ui_payload_rejects_singleton_multidisorder_step() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("multidisorder", "host+1")];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("singleton multidisorder");
-
-    assert!(err.to_string().contains("multidisorder must declare at least two markers"));
-}
-
-#[test]
-fn ui_payload_rejects_mixed_multidisorder_chain() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("multidisorder", "host+1"), tcp_step("split", "midsld")];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("mixed multidisorder");
-
-    assert!(err.to_string().contains("multidisorder must be the only tcp send step family"));
-}
-
-#[test]
-fn ui_payload_parses_ipfrag_steps_and_udp_split_bytes() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("tlsrec", "extlen"), tcp_step("ipfrag2", "host+2")];
-    ui.chains.udp_steps = vec![ProxyUiUdpChainStep {
-        kind: "ipfrag2_udp".to_string(),
-        count: 0,
-        split_bytes: 8,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-    }];
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    let group = &config.groups[0];
-
-    assert_eq!(group.actions.tcp_chain.len(), 2);
-    assert_eq!(group.actions.tcp_chain[0].kind(), TcpChainStepKind::TlsRec);
-    assert_eq!(group.actions.tcp_chain[1].kind(), TcpChainStepKind::IpFrag2);
-    assert_eq!(group.actions.tcp_chain[1].offset().base, OffsetBase::Host);
-    assert_eq!(group.actions.tcp_chain[1].offset().delta, 2);
-    assert_eq!(group.actions.udp_chain.len(), 1);
-    assert_eq!(group.actions.udp_chain[0].kind, UdpChainStepKind::IpFrag2Udp);
-    assert_eq!(group.actions.udp_chain[0].count, 0);
-    assert_eq!(group.actions.udp_chain[0].split_bytes, 8);
-}
-
-#[test]
-fn ui_payload_parses_tcp_rotation_policy_defaults() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("tlsrec", "extlen"), tcp_step("split", "host+2")];
-    ui.chains.tcp_rotation = Some(ProxyUiTcpRotationConfig {
-        candidates: vec![
-            ProxyUiTcpRotationCandidate {
-                tcp_steps: vec![
-                    tcp_step("tlsrec", "extlen"),
-                    ProxyUiTcpChainStep {
-                        kind: "hostfake".to_string(),
-                        marker: "endhost+8".to_string(),
-                        midhost_marker: "midsld".to_string(),
-                        fake_host_template: "googlevideo.com".to_string(),
-                        fake_order: String::new(),
-                        fake_seq_mode: String::new(),
-                        tcp_flags_set: String::new(),
-                        tcp_flags_unset: String::new(),
-                        tcp_flags_orig_set: String::new(),
-                        tcp_flags_orig_unset: String::new(),
-                        overlap_size: 0,
-                        fake_mode: String::new(),
-                        fragment_count: 0,
-                        min_fragment_size: 0,
-                        max_fragment_size: 0,
-                        inter_segment_delay_ms: 0,
-                        activation_filter: None,
-                        ipv6_extension_profile: "none".to_string(),
-                        random_fake_host: false,
-                    },
-                    tcp_step("split", "midsld"),
-                ],
-            },
-            ProxyUiTcpRotationCandidate { tcp_steps: vec![tcp_step("split", "host+2")] },
-        ],
-        ..ProxyUiTcpRotationConfig::default()
-    });
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    let rotation = config.groups[0].actions.rotation_policy.as_ref().expect("rotation policy");
-
-    assert_eq!(rotation.fails, 3);
-    assert_eq!(rotation.retrans, 3);
-    assert_eq!(rotation.seq, 65_536);
-    assert_eq!(rotation.rst, 1);
-    assert_eq!(rotation.time_secs, 60);
-    assert_eq!(rotation.candidates.len(), 2);
-    assert_eq!(rotation.candidates[0].tcp_chain[0].kind(), TcpChainStepKind::TlsRec);
-    assert_eq!(rotation.candidates[0].tcp_chain[1].kind(), TcpChainStepKind::HostFake);
-    assert_eq!(rotation.candidates[0].tcp_chain[2].kind(), TcpChainStepKind::Split);
-    assert_eq!(rotation.candidates[1].tcp_chain[0].kind(), TcpChainStepKind::Split);
-}
-
-#[test]
-fn ui_payload_rejects_empty_tcp_rotation_candidates() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_rotation =
-        Some(ProxyUiTcpRotationConfig { candidates: Vec::new(), ..ProxyUiTcpRotationConfig::default() });
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("empty rotation");
-
-    assert!(err.to_string().contains("chains.tcpRotation must declare at least one candidate"));
-}
-
-#[test]
-fn ui_payload_rejects_malformed_tcp_rotation_candidate_chain() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_rotation = Some(ProxyUiTcpRotationConfig {
-        candidates: vec![ProxyUiTcpRotationCandidate {
-            tcp_steps: vec![seqovl_step("host+1", 12, "profile"), seqovl_step("midsld", 16, "rand")],
-        }],
-        ..ProxyUiTcpRotationConfig::default()
-    });
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("invalid rotation candidate");
-
-    assert!(err.to_string().contains("seqovl must appear at most once per tcp chain"));
-}
-
-#[test]
-fn ui_payload_treats_fake_approx_steps_as_fake_payload_consumers() {
-    for kind in ["fakedsplit", "fakeddisorder"] {
-        let mut ui = minimal_ui();
-        ui.chains.tcp_steps = vec![tcp_step(kind, "host+1")];
-        ui.fake_packets.fake_tls_use_original = true;
-        ui.fake_packets.fake_tls_dup_session_id = true;
-        ui.fake_packets.fake_offset_marker = "host+1".to_string();
-
-        let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-        let group = &config.groups[0];
-
-        let expected_mode = if kind == "fakedsplit" { DesyncMode::Fake } else { DesyncMode::Disorder };
-
-        assert_eq!(group.actions.tcp_chain[0].kind().as_mode(), Some(expected_mode));
-        assert_eq!(group.actions.fake_offset.map(|offset| offset.delta), Some(1));
-        assert_ne!(group.actions.fake_mod & FM_ORIG, 0);
-        assert_ne!(group.actions.fake_mod & FM_DUPSID, 0);
-    }
-}
-
-#[test]
-fn ui_payload_rejects_non_terminal_fake_approx_step() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps.push(tcp_step("fakedsplit", "host+1"));
-    ui.chains.tcp_steps.push(tcp_step("split", "endhost"));
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("non-terminal fakedsplit");
-
-    assert!(err.to_string().contains("fakedsplit"));
-}
-
-#[test]
-fn ui_payload_parses_fake_payload_profiles() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.http_fake_profile = "cloudflare_get".to_string();
-    ui.fake_packets.tls_fake_profile = "google_chrome".to_string();
-    ui.fake_packets.udp_fake_profile = "dns_query".to_string();
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert_eq!(config.groups[0].actions.http_fake_profile, HttpFakeProfile::CloudflareGet);
-    assert_eq!(config.groups[0].actions.tls_fake_profile, TlsFakeProfile::GoogleChrome);
-    assert_eq!(config.groups[0].actions.udp_fake_profile, UdpFakeProfile::DnsQuery);
-}
-
-#[test]
-fn ui_payload_maps_fake_tls_source_and_secondary_profile() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.fake_tls_source = "captured_client_hello".to_string();
-    ui.fake_packets.fake_tls_secondary_profile = "google_chrome".to_string();
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert_eq!(config.groups[0].actions.fake_tls_source, FakePacketSource::CapturedClientHello);
-    assert_eq!(config.groups[0].actions.fake_tls_secondary_profile, Some(TlsFakeProfile::GoogleChrome));
-}
-
-#[test]
-fn ui_payload_maps_fake_tcp_timestamp_settings() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.fake_tcp_timestamp_enabled = true;
-    ui.fake_packets.fake_tcp_timestamp_delta_ticks = -77;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert!(config.groups[0].actions.fake_tcp_timestamp_enabled);
-    assert_eq!(config.groups[0].actions.fake_tcp_timestamp_delta_ticks, -77);
-}
-
-#[test]
-fn ui_payload_maps_ip_id_mode_seqgroup() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.ip_id_mode = "seqgroup".to_string();
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert_eq!(config.groups[0].actions.ip_id_mode, Some(ripdpi_config::IpIdMode::SeqGroup));
-}
-
-#[test]
-fn ui_payload_rejects_invalid_ip_id_mode() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.ip_id_mode = "bogus".to_string();
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("invalid ip id mode");
-
-    assert!(err.to_string().contains("fakePackets.ipIdMode"));
-}
-
-#[test]
-fn ui_payload_parses_fake_order_and_seq_mode() {
-    let mut ui = minimal_ui();
-    let mut step = tcp_step("fakedsplit", "host+1");
-    step.fake_order = "2".to_string();
-    step.fake_seq_mode = "sequential".to_string();
-    ui.chains.tcp_steps = vec![tcp_step("tlsrec", "extlen"), step];
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    let parsed = &config.groups[0].actions.tcp_chain[1];
-
-    assert_eq!(parsed.kind(), TcpChainStepKind::FakeSplit);
-    assert_eq!(parsed.fake_ordering().order, FakeOrder::RealFakeRealFake);
-    assert_eq!(parsed.fake_ordering().seq_mode, FakeSeqMode::Sequential);
-}
-
-#[test]
-fn ui_payload_rejects_fake_order_on_unsupported_step_kind() {
-    let mut ui = minimal_ui();
-    let mut step = tcp_step("split", "host+1");
-    step.fake_order = "1".to_string();
-    ui.chains.tcp_steps = vec![step];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("unsupported fake ordering");
-
-    assert!(err.to_string().contains("must not declare fake ordering fields"));
-}
-
-#[test]
-fn ui_payload_rejects_hostfake_non_default_order_without_midhost() {
-    let mut ui = minimal_ui();
-    let mut step = tcp_step("hostfake", "endhost+8");
-    step.fake_order = "1".to_string();
-    ui.chains.tcp_steps = vec![tcp_step("tlsrec", "extlen"), step];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("hostfake altorder without midhost");
-
-    assert!(err.to_string().contains("hostfake fakeOrder requires midhostMarker"));
-}
-
-#[test]
-fn ui_payload_maps_extended_http_parser_evasions_into_mod_http() {
-    let mut ui = minimal_ui();
-    ui.parser_evasions.host_mixed_case = true;
-    ui.parser_evasions.domain_mixed_case = true;
-    ui.parser_evasions.host_remove_spaces = true;
-    ui.parser_evasions.http_method_eol = true;
-    ui.parser_evasions.http_unix_eol = true;
-    ui.parser_evasions.http_host_extra_space = true;
-    ui.parser_evasions.http_host_tab = true;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert_eq!(
-        config.groups[0].actions.mod_http,
-        MH_HMIX | MH_DMIX | MH_SPACE | MH_METHODEOL | MH_UNIXEOL | MH_HOSTEXTRASPACE | MH_HOSTTAB
-    );
-}
-
-#[test]
-fn ui_payload_rejects_unknown_ipv6_extension_profile() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![ProxyUiTcpChainStep {
-        kind: "ipfrag2".to_string(),
-        marker: "host+2".to_string(),
-        midhost_marker: String::new(),
-        fake_host_template: String::new(),
-        fake_order: String::new(),
-        fake_seq_mode: String::new(),
-        tcp_flags_set: String::new(),
-        tcp_flags_unset: String::new(),
-        tcp_flags_orig_set: String::new(),
-        tcp_flags_orig_unset: String::new(),
-        overlap_size: 0,
-        fake_mode: String::new(),
-        fragment_count: 0,
-        min_fragment_size: 0,
-        max_fragment_size: 0,
-        inter_segment_delay_ms: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "bogus".to_string(),
-        random_fake_host: false,
-    }];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("invalid ipv6 extension profile");
-
-    assert!(err.to_string().contains("Unsupported ipv6ExtensionProfile"));
-}
-
-#[test]
-fn legacy_flat_ui_json_is_rejected() {
-    let err = parse_proxy_config_json(
-        &serde_json::json!({
-            "kind": "ui",
-            "ip": "127.0.0.1",
-            "port": 1080,
-            "desyncMethod": "disorder",
-        })
-        .to_string(),
-    )
-    .expect_err("legacy flat ui config should be rejected");
-
-    assert!(err.to_string().contains("Legacy flat UI config JSON is not supported"));
-}
-
-#[test]
-fn grouped_ui_json_requires_at_least_one_nested_section() {
-    let err = parse_proxy_config_json(&serde_json::json!({ "kind": "ui" }).to_string())
-        .expect_err("ui payload without grouped sections should be rejected");
-
-    assert!(err.to_string().contains("at least one nested section"));
-}
-
-// --- schemaVersion envelope tests ---
-
-/// Serializes `ui_payload(minimal_ui())` to a current-version JSON object.
-fn ui_payload_json_object() -> serde_json::Map<String, serde_json::Value> {
-    let mut payload = serde_json::to_value(ui_payload(minimal_ui())).expect("serialize ui payload");
-    let object = payload.as_object_mut().expect("ui payload object");
-    object.clone()
-}
-
-#[test]
-fn payload_without_schema_version_is_rejected() {
-    let mut object = ui_payload_json_object();
-    object.remove("schemaVersion");
-
-    let err = parse_proxy_config_json(&serde_json::Value::Object(object).to_string())
-        .expect_err("payload without schemaVersion should be rejected");
-
-    assert!(err.to_string().contains("schemaVersion"), "error should name schemaVersion: {err}");
-}
-
-#[test]
-fn payload_with_explicit_schema_version_two_parses() {
-    let mut object = ui_payload_json_object();
-    object.insert("schemaVersion".to_string(), serde_json::json!(2));
-
-    let parsed = parse_proxy_config_json(&serde_json::Value::Object(object).to_string())
-        .expect("payload with schemaVersion 2 should parse");
-
-    assert_eq!(parsed.schema_version(), 2);
-}
-
-#[test]
-fn payload_with_old_schema_version_is_rejected() {
-    let mut object = ui_payload_json_object();
-    object.insert("schemaVersion".to_string(), serde_json::json!(1));
-
-    let err = parse_proxy_config_json(&serde_json::Value::Object(object).to_string())
-        .expect_err("payload with schemaVersion 1 should be rejected");
-
-    assert_eq!(err, ProxyConfigError::UnsupportedSchemaVersion { found: 1 });
-    assert!(
-        err.to_string().contains("unsupported native config schemaVersion 1"),
-        "error should name the found version, got: {err}"
-    );
-}
-
-#[test]
-fn payload_with_future_schema_version_is_rejected() {
-    let mut object = ui_payload_json_object();
-    object.insert("schemaVersion".to_string(), serde_json::json!(3));
-
-    let err = parse_proxy_config_json(&serde_json::Value::Object(object).to_string())
-        .expect_err("payload with schemaVersion 3 should be rejected");
-
-    assert_eq!(err, ProxyConfigError::UnsupportedSchemaVersion { found: 3 });
-}
-
-#[test]
-fn runtime_config_envelope_rejects_unsupported_schema_version() {
-    // A payload constructed outside `parse_proxy_config_json` is still validated
-    // at the conversion entry point.
-    let payload = ProxyConfigPayload::CommandLine {
-        args: vec!["ripdpi".to_string(), "--split".to_string(), "host+1".to_string()],
-        host_autolearn_store_path: None,
-        runtime_context: None,
-        log_context: None,
-        session_overrides: None,
-        schema_version: 3,
-    };
-
-    let err = runtime_config_envelope_from_payload(payload).expect_err("schemaVersion 3 should be rejected");
-
-    assert_eq!(err, ProxyConfigError::UnsupportedSchemaVersion { found: 3 });
-}
-
-#[test]
-fn command_line_payload_requires_runnable_config() {
-    let err = runtime_config_from_payload(ProxyConfigPayload::CommandLine {
-        args: vec!["ripdpi".to_string(), "--help".to_string()],
-        host_autolearn_store_path: None,
-        runtime_context: None,
-        log_context: None,
-        session_overrides: None,
-        schema_version: 2,
-    })
-    .expect_err("help should not produce runnable config");
-
-    assert!(err.to_string().contains("runnable"));
-}
-
-#[test]
-fn command_line_session_overrides_apply_ephemeral_port_and_auth_token() {
-    let envelope = runtime_config_envelope_from_payload(ProxyConfigPayload::CommandLine {
-        args: vec!["ripdpi".to_string(), "--port".to_string(), "1081".to_string()],
-        host_autolearn_store_path: None,
-        runtime_context: None,
-        log_context: None,
-        session_overrides: Some(ProxySessionOverrides {
-            listen_port_override: Some(0),
-            auth_token: Some("alpha-123".to_string()),
-        }),
-        schema_version: 2,
-    })
-    .expect("runtime config envelope");
-
-    assert_eq!(envelope.config.network.listen.listen_port, 0);
-    assert_eq!(envelope.config.network.listen.auth_token.as_deref(), Some("alpha-123"));
-}
-
-#[test]
-fn kotlin_ui_json_session_overrides_apply_ephemeral_port_and_auth_token() {
-    let mut payload = serde_json::to_value(minimal_ui()).expect("serialize minimal ui");
-    let object = payload.as_object_mut().expect("minimal ui object");
-    object.insert("kind".to_string(), serde_json::json!("ui"));
-    object.insert("schemaVersion".to_string(), serde_json::json!(2));
-    object.insert("strategyPreset".to_string(), serde_json::json!("ripdpi_default"));
-    object.insert(
-        "runtimeContext".to_string(),
-        serde_json::json!({
-            "protectPath": "/tmp/protect.sock",
-        }),
-    );
-    object.insert(
-        "logContext".to_string(),
-        serde_json::json!({
-            "runtimeId": "runtime-1",
-            "mode": "VPN",
-        }),
-    );
-    object.insert(
-        "sessionOverrides".to_string(),
-        serde_json::json!({
-            "listenPortOverride": 0,
-            "authToken": "alpha-123",
-        }),
-    );
-
-    let parsed = parse_proxy_config_json(&payload.to_string()).expect("parse kotlin ui json");
-    let envelope = runtime_config_envelope_from_payload(parsed).expect("runtime config envelope");
-
-    assert_eq!(envelope.config.network.listen.listen_port, 0);
-    assert_eq!(envelope.config.network.listen.auth_token.as_deref(), Some("alpha-123"));
-    assert_eq!(envelope.runtime_context.and_then(|context| context.protect_path).as_deref(), Some("/tmp/protect.sock"),);
-    assert_eq!(envelope.log_context.and_then(|context| context.runtime_id).as_deref(), Some("runtime-1"),);
-}
-
-#[test]
-fn invalid_session_override_listen_port_is_rejected() {
-    let err = runtime_config_envelope_from_payload(ProxyConfigPayload::CommandLine {
-        args: vec!["ripdpi".to_string(), "--port".to_string(), "1081".to_string()],
-        host_autolearn_store_path: None,
-        runtime_context: None,
-        log_context: None,
-        session_overrides: Some(ProxySessionOverrides {
-            listen_port_override: Some(-1),
-            auth_token: Some("alpha-123".to_string()),
-        }),
-        schema_version: 2,
-    })
-    .expect_err("invalid listen port override");
-
-    assert!(err.to_string().contains("listenPortOverride"));
-}
-
-#[test]
-fn runtime_context_sanitizes_direct_path_capabilities() {
-    let envelope = runtime_config_envelope_from_payload(ProxyConfigPayload::CommandLine {
-        args: vec!["ripdpi".to_string(), "--split".to_string(), "host+1".to_string()],
-        host_autolearn_store_path: None,
-        runtime_context: Some(ProxyRuntimeContext {
-            encrypted_dns: None,
-            protect_path: None,
-            preferred_edges: std::collections::BTreeMap::default(),
-            direct_path_capabilities: vec![
-                ProxyDirectPathCapability {
-                    authority: " Example.org:443. ".to_string(),
-                    quic_usable: Some(false),
-                    udp_usable: Some(false),
-                    fallback_required: Some(true),
-                    repeated_handshake_failure_class: Some(" tcp_reset ".to_string()),
-                    transport_policy_version: -1,
-                    ip_set_digest: " 203.0.113.10,203.0.113.11 ".to_string(),
-                    dns_classification: Some(" ech_capable ".to_string()),
-                    quic_mode: " soft_disable ".to_string(),
-                    preferred_stack: " h2 ".to_string(),
-                    dns_mode: " doh_primary ".to_string(),
-                    tcp_family: " rec_pre_sni ".to_string(),
-                    outcome: " no_direct_solution ".to_string(),
-                    transport_class: Some(" sni_tls_suspect ".to_string()),
-                    reason_code: Some(" tcp_post_client_hello_failure ".to_string()),
-                    cooldown_until: Some(-1),
-                    updated_at: -10,
-                },
-                ProxyDirectPathCapability {
-                    authority: "   ".to_string(),
-                    quic_usable: None,
-                    udp_usable: None,
-                    fallback_required: None,
-                    repeated_handshake_failure_class: None,
-                    transport_policy_version: 0,
-                    ip_set_digest: String::new(),
-                    dns_classification: None,
-                    quic_mode: "ALLOW".to_string(),
-                    preferred_stack: "H3".to_string(),
-                    dns_mode: "SYSTEM".to_string(),
-                    tcp_family: "NONE".to_string(),
-                    outcome: "TRANSPARENT_OK".to_string(),
-                    transport_class: None,
-                    reason_code: None,
-                    cooldown_until: None,
-                    updated_at: 0,
-                },
-            ],
-            morph_policy: Some(ProxyMorphPolicy {
-                id: " balanced ".to_string(),
-                first_flight_size_min: -10,
-                first_flight_size_max: 700,
-                padding_envelope_min: -1,
-                padding_envelope_max: 80,
-                entropy_target_permil: -50,
-                tcp_burst_cadence_ms: vec![0, -5, 12],
-                tls_burst_cadence_ms: vec![8, -2],
-                quic_burst_profile: " Compat_Burst ".to_string(),
-                fake_packet_shape_profile: " Compat_Default ".to_string(),
-            }),
-            connection_concurrency: None,
-        }),
-        log_context: None,
-        session_overrides: None,
-        schema_version: 2,
-    })
-    .expect("runtime config envelope");
-
-    let mut runtime_context = envelope.runtime_context.expect("runtime context");
-    let capability = runtime_context.direct_path_capabilities.pop().expect("capability");
-    let morph_policy = runtime_context.morph_policy.expect("morph policy");
-
-    assert_eq!(capability.authority, "example.org:443");
-    assert_eq!(capability.quic_usable, Some(false));
-    assert_eq!(capability.udp_usable, Some(false));
-    assert_eq!(capability.fallback_required, Some(true));
-    assert_eq!(capability.repeated_handshake_failure_class.as_deref(), Some("tcp_reset"));
-    assert_eq!(capability.transport_policy_version, 0);
-    assert_eq!(capability.ip_set_digest, "203.0.113.10,203.0.113.11");
-    assert_eq!(capability.dns_classification.as_deref(), Some("ECH_CAPABLE"));
-    assert_eq!(capability.quic_mode, "SOFT_DISABLE");
-    assert_eq!(capability.preferred_stack, "H2");
-    assert_eq!(capability.dns_mode, "DOH_PRIMARY");
-    assert_eq!(capability.tcp_family, "REC_PRE_SNI");
-    assert_eq!(capability.outcome, "NO_DIRECT_SOLUTION");
-    assert_eq!(capability.transport_class.as_deref(), Some("SNI_TLS_SUSPECT"));
-    assert_eq!(capability.reason_code.as_deref(), Some("TCP_POST_CLIENT_HELLO_FAILURE"));
-    assert_eq!(capability.cooldown_until, None);
-    assert_eq!(capability.updated_at, 0);
-    assert_eq!(morph_policy.id, "balanced");
-    assert_eq!(morph_policy.first_flight_size_min, 0);
-    assert_eq!(morph_policy.first_flight_size_max, 700);
-    assert_eq!(morph_policy.padding_envelope_min, 0);
-    assert_eq!(morph_policy.padding_envelope_max, 80);
-    assert_eq!(morph_policy.entropy_target_permil, 0);
-    assert_eq!(morph_policy.tcp_burst_cadence_ms, vec![0, 0, 12]);
-    assert_eq!(morph_policy.tls_burst_cadence_ms, vec![8, 0]);
-    assert_eq!(morph_policy.quic_burst_profile, "compat_burst");
-    assert_eq!(morph_policy.fake_packet_shape_profile, "compat_default");
-}
-
-#[test]
-fn invalid_quic_fake_profile_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.quic.fake_profile = "bogus".to_string();
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("invalid quic profile");
-
-    assert!(err.to_string().contains("quicFakeProfile"));
-}
-
-#[test]
-fn invalid_http_fake_profile_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.http_fake_profile = "bogus".to_string();
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("invalid http profile");
-
-    assert!(err.to_string().contains("httpFakeProfile"));
-}
-
-#[test]
-fn adaptive_hostfake_marker_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps.push(tcp_step("hostfake", "auto(host)"));
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("adaptive hostfake marker");
-
-    assert!(err.to_string().contains("hostfake"));
-}
-
-#[test]
-fn adaptive_hostfake_midhost_marker_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps.push(ProxyUiTcpChainStep {
-        kind: "hostfake".to_string(),
-        marker: "endhost+8".to_string(),
-        midhost_marker: "auto(midsld)".to_string(),
-        fake_host_template: String::new(),
-        fake_order: String::new(),
-        fake_seq_mode: String::new(),
-        tcp_flags_set: String::new(),
-        tcp_flags_unset: String::new(),
-        tcp_flags_orig_set: String::new(),
-        tcp_flags_orig_unset: String::new(),
-        overlap_size: 0,
-        fake_mode: String::new(),
-        fragment_count: 0,
-        min_fragment_size: 0,
-        max_fragment_size: 0,
-        inter_segment_delay_ms: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-        random_fake_host: false,
-    });
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("adaptive hostfake midhost");
-
-    assert!(err.to_string().contains("midhostMarker"));
-}
-
-#[test]
-fn adaptive_fake_offset_marker_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("fake", "host+1")];
-    ui.fake_packets.fake_offset_marker = "auto(host)".to_string();
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("adaptive fake offset");
-
-    assert!(err.to_string().contains("fakeOffsetMarker"));
-}
-
-#[test]
-fn empty_fake_offset_marker_uses_compat_default_offset() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("fake", "host+1")];
-    ui.fake_packets.fake_offset_marker.clear();
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("fake offset fallback");
-
-    assert_eq!(config.groups[0].actions.fake_offset, Some(OffsetExpr::absolute(0)));
-}
-
-#[test]
-fn ui_payload_rejects_non_terminal_ipfrag2_step() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps =
-        vec![tcp_step("tlsrec", "extlen"), tcp_step("ipfrag2", "host+2"), tcp_step("split", "endhost")];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("non-terminal ipfrag2");
-
-    assert!(err.to_string().contains("ipfrag2 must be the only tcp send step"));
-}
-
-#[test]
-fn ui_payload_rejects_ipfrag2_udp_with_count() {
-    let mut ui = minimal_ui();
-    ui.chains.udp_steps = vec![ProxyUiUdpChainStep {
-        kind: "ipfrag2_udp".to_string(),
-        count: 1,
-        split_bytes: 8,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-    }];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("ipfrag2_udp count");
-
-    assert!(err.to_string().contains("must not declare count"));
-}
-
-#[test]
-fn ui_payload_rejects_ipfrag2_udp_without_positive_split_bytes() {
-    let mut ui = minimal_ui();
-    ui.chains.udp_steps = vec![ProxyUiUdpChainStep {
-        kind: "ipfrag2_udp".to_string(),
-        count: 0,
-        split_bytes: 0,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-    }];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("ipfrag2_udp splitBytes");
-
-    assert!(err.to_string().contains("must declare positive splitBytes"));
-}
-
-#[test]
-fn ui_payload_rejects_split_bytes_for_non_ipfrag_udp_steps() {
-    let mut ui = minimal_ui();
-    ui.chains.udp_steps = vec![ProxyUiUdpChainStep {
-        kind: "fake_burst".to_string(),
-        count: 2,
-        split_bytes: 8,
-        activation_filter: None,
-        ipv6_extension_profile: "none".to_string(),
-    }];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("fake_burst splitBytes");
-
-    assert!(err.to_string().contains("splitBytes is only supported"));
-}
-
-#[test]
-fn ui_payload_rejects_mixed_ipfrag2_udp_chain() {
-    let mut ui = minimal_ui();
-    ui.chains.udp_steps = vec![
-        ProxyUiUdpChainStep {
-            kind: "ipfrag2_udp".to_string(),
-            count: 0,
-            split_bytes: 8,
-            activation_filter: None,
-            ipv6_extension_profile: "none".to_string(),
-        },
-        udp_step("fake_burst", 1),
-    ];
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("mixed ipfrag2_udp chain");
-
-    assert!(err.to_string().contains("ipfrag2_udp must be the only udp chain step"));
-}
-
-#[test]
-fn ech_fake_offset_marker_is_rejected() {
-    for marker in ["echext", "echext+4"] {
-        let mut ui = minimal_ui();
-        ui.chains.tcp_steps = vec![tcp_step("fake", "host+1")];
-        ui.fake_packets.fake_offset_marker = marker.to_string();
-
-        let err = runtime_config_from_payload(ui_payload(ui)).expect_err("ech fake offset");
-
-        assert!(err.to_string().contains("fakeOffsetMarker"), "{marker} should reference fakeOffsetMarker");
-    }
-}
-
-#[test]
-fn ui_payload_maps_adaptive_fake_ttl_and_fallback() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.fake_ttl = 11;
-    ui.fake_packets.adaptive_fake_ttl_enabled = true;
-    ui.fake_packets.adaptive_fake_ttl_delta = -1;
-    ui.fake_packets.adaptive_fake_ttl_min = 3;
-    ui.fake_packets.adaptive_fake_ttl_max = 12;
-    ui.fake_packets.adaptive_fake_ttl_fallback = 9;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("adaptive fake ttl config");
-    let group = &config.groups[0];
-
-    assert_eq!(group.actions.auto_ttl, Some(AutoTtlConfig { delta: -1, min_ttl: 3, max_ttl: 12 }),);
-    assert_eq!(group.actions.ttl, Some(9));
-}
-
-#[test]
-fn ui_payload_rejects_invalid_adaptive_fake_ttl_window() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.adaptive_fake_ttl_enabled = true;
-    ui.fake_packets.adaptive_fake_ttl_min = 12;
-    ui.fake_packets.adaptive_fake_ttl_max = 3;
-
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("invalid adaptive ttl window");
-
-    assert!(err.to_string().contains("adaptive fake TTL window"));
-}
-
-#[test]
-fn ui_payload_uses_fixed_fake_ttl_when_adaptive_is_disabled() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.fake_ttl = 13;
-    ui.fake_packets.adaptive_fake_ttl_enabled = false;
-    ui.fake_packets.adaptive_fake_ttl_fallback = 5;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("fixed fake ttl config");
-    let group = &config.groups[0];
-
-    assert_eq!(group.actions.auto_ttl, None);
-    assert_eq!(group.actions.ttl, Some(13));
-}
-
-#[test]
-fn known_preset_applies_without_error() {
-    let base = minimal_ui();
-    let mut cfg = base.clone();
-    crate::presets::apply_preset("russia_rostelecom", &mut cfg).unwrap();
-    assert!(cfg.protocols.desync_https, "rostelecom preset should enable HTTPS desync");
-    assert!(cfg.fake_packets.adaptive_fake_ttl_enabled, "rostelecom preset should enable adaptive TTL");
-}
-
-#[test]
-fn unknown_preset_returns_error() {
-    let mut cfg = minimal_ui();
-    let result = crate::presets::apply_preset("not_a_real_preset", &mut cfg);
-    assert!(result.is_err());
-}
-
-#[test]
-fn russia_mgts_preset_produces_valid_runtime_config() {
-    let mut cfg = minimal_ui();
-    crate::presets::apply_preset("russia_mgts", &mut cfg).unwrap();
-    runtime_config_from_ui(cfg).unwrap();
-}
-
-#[test]
-fn preset_field_in_ui_config_round_trips_json() {
-    let cfg = minimal_ui();
-    let payload = ProxyConfigPayload::Ui {
-        strategy_preset: Some("ripdpi_default".to_string()),
-        config: cfg,
-        runtime_context: None,
-        log_context: None,
-        session_overrides: None,
-        schema_version: 2,
-    };
-    let json = serde_json::to_string(&payload).unwrap();
-    let decoded: ProxyConfigPayload = serde_json::from_str(&json).unwrap();
-    match decoded {
-        ProxyConfigPayload::Ui { strategy_preset, .. } => {
-            assert_eq!(strategy_preset.as_deref(), Some("ripdpi_default"));
-        }
-        other @ ProxyConfigPayload::CommandLine { .. } => panic!("expected ui payload, got {other:?}"),
-    }
-}
-
-// --- NetworkSnapshot serde tests ---
-
-#[test]
-fn network_snapshot_deserializes_from_empty_json() {
-    let snapshot: NetworkSnapshot = serde_json::from_str("{}").expect("deserialize");
-    assert_eq!(snapshot.transport, "");
-    assert!(!snapshot.validated);
-    assert!(!snapshot.captive_portal);
-    assert!(!snapshot.metered);
-    assert_eq!(snapshot.private_dns_mode, "");
-    assert!(snapshot.dns_servers.is_empty());
-    assert!(snapshot.cellular.is_none());
-    assert!(snapshot.wifi.is_none());
-    assert!(snapshot.mtu.is_none());
-    assert_eq!(snapshot.traffic_tx_bytes, 0);
-    assert_eq!(snapshot.traffic_rx_bytes, 0);
-    assert_eq!(snapshot.captured_at_ms, 0);
-}
-
-#[test]
-fn network_snapshot_round_trips_wifi_snapshot() {
-    let snapshot = NetworkSnapshot {
-        transport: "wifi".to_string(),
-        validated: true,
-        captive_portal: false,
-        metered: false,
-        private_dns_mode: "system".to_string(),
-        dns_servers: vec!["8.8.8.8".to_string(), "8.8.4.4".to_string()],
-        cellular: None,
-        wifi: Some(WifiSnapshot {
-            frequency_band: "5ghz".to_string(),
-            ssid_hash: "abc123def456".to_string(),
-            frequency_mhz: Some(5180),
-            rssi_dbm: Some(-58),
-            link_speed_mbps: Some(866),
-            rx_link_speed_mbps: Some(780),
-            tx_link_speed_mbps: Some(720),
-            channel_width: "80 MHz".to_string(),
-            wifi_standard: "802.11ax".to_string(),
-        }),
-        mtu: Some(1500),
-        traffic_tx_bytes: 1_234_567,
-        traffic_rx_bytes: 9_876_543,
-        captured_at_ms: 1_700_000_000_000,
-        vpn_service_was_active: false,
-    };
-    let json = serde_json::to_string(&snapshot).expect("serialize");
-    let decoded: NetworkSnapshot = serde_json::from_str(&json).expect("deserialize");
-    assert_eq!(decoded, snapshot);
-}
-
-#[test]
-fn network_snapshot_round_trips_cellular_snapshot() {
-    let snapshot = NetworkSnapshot {
-        transport: "cellular".to_string(),
-        validated: true,
-        captive_portal: false,
-        metered: true,
-        private_dns_mode: "system".to_string(),
-        dns_servers: vec!["10.0.0.1".to_string()],
-        cellular: Some(CellularSnapshot {
-            generation: "4g".to_string(),
-            roaming: false,
-            operator_code: "25001".to_string(),
-            data_network_type: "LTE".to_string(),
-            service_state: "in_service".to_string(),
-            carrier_id: Some(42),
-            signal_level: Some(4),
-            signal_dbm: Some(-95),
-        }),
-        wifi: None,
-        mtu: Some(1420),
-        traffic_tx_bytes: 0,
-        traffic_rx_bytes: 0,
-        captured_at_ms: 1_700_000_000_000,
-        vpn_service_was_active: false,
-    };
-    let json = serde_json::to_string(&snapshot).expect("serialize");
-    let decoded: NetworkSnapshot = serde_json::from_str(&json).expect("deserialize");
-    assert_eq!(decoded, snapshot);
-}
-
-#[test]
-fn network_snapshot_ignores_unknown_fields() {
-    let json = r#"{
-        "transport": "wifi",
-        "validated": true,
-        "mtu": 1500,
-        "futureField": "some_value",
-        "anotherNewField": 42
-    }"#;
-    let snapshot: NetworkSnapshot = serde_json::from_str(json).expect("deserialize with unknown fields");
-    assert_eq!(snapshot.transport, "wifi");
-    assert!(snapshot.validated);
-    assert_eq!(snapshot.mtu, Some(1500));
-}
-
-#[test]
-fn network_snapshot_uses_camel_case_keys() {
-    let snapshot = NetworkSnapshot {
-        transport: "cellular".to_string(),
-        metered: true,
-        captive_portal: true,
-        private_dns_mode: "dns.example.com".to_string(),
-        cellular: Some(CellularSnapshot {
-            data_network_type: "NR".to_string(),
-            service_state: "in_service".to_string(),
-            carrier_id: Some(7),
-            signal_level: Some(4),
-            signal_dbm: Some(-93),
-            ..CellularSnapshot::default()
-        }),
-        wifi: Some(WifiSnapshot {
-            frequency_mhz: Some(5955),
-            channel_width: "80 MHz".to_string(),
-            wifi_standard: "802.11be".to_string(),
-            ..WifiSnapshot::default()
-        }),
-        mtu: Some(1280),
-        ..NetworkSnapshot::default()
-    };
-    let json = serde_json::to_string(&snapshot).expect("serialize");
-    assert!(json.contains("\"captivePortal\""), "expected camelCase key in: {json}");
-    assert!(json.contains("\"mtu\""), "expected mtu key in: {json}");
-    assert!(json.contains("\"privateDnsMode\""), "expected camelCase key in: {json}");
-    assert!(json.contains("\"dataNetworkType\""), "expected dataNetworkType key in: {json}");
-    assert!(json.contains("\"signalLevel\""), "expected signalLevel key in: {json}");
-    assert!(json.contains("\"frequencyMhz\""), "expected frequencyMhz key in: {json}");
-    assert!(json.contains("\"wifiStandard\""), "expected wifiStandard key in: {json}");
-}
-
-// --- WsTunnelMode conversion tests ---
-
-#[test]
-fn ws_tunnel_mode_fallback() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.mode = Some("fallback".to_string());
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_mode, WsTunnelMode::Fallback);
-}
-
-#[test]
-fn ws_tunnel_mode_always() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.mode = Some("always".to_string());
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_mode, WsTunnelMode::Always);
-}
-
-#[test]
-fn ws_tunnel_mode_off() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.mode = Some("off".to_string());
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_mode, WsTunnelMode::Off);
-}
-
-#[test]
-fn ws_tunnel_mode_unknown_string_maps_to_off() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.mode = Some("unknown_string".to_string());
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_mode, WsTunnelMode::Off);
-}
-
-#[test]
-fn ws_tunnel_allow_insecure_sni_defaults_to_false() {
-    let config = runtime_config_from_ui(minimal_ui()).expect("runtime config");
-    assert!(!config.adaptive.ws_tunnel_allow_insecure_sni);
-}
-
-#[test]
-fn ws_tunnel_allow_insecure_sni_maps_from_ui() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.fake_sni = Some("yandex.ru".to_string());
-    ui.ws_tunnel.allow_insecure_sni = true;
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_fake_sni.as_deref(), Some("yandex.ru"));
-    assert!(config.adaptive.ws_tunnel_allow_insecure_sni);
-}
-
-#[test]
-fn ws_tunnel_mode_none_enabled_true_maps_to_always() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.mode = None;
-    ui.ws_tunnel.enabled = true;
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_mode, WsTunnelMode::Always);
-}
-
-#[test]
-fn ws_tunnel_mode_none_enabled_false_maps_to_off() {
-    let mut ui = minimal_ui();
-    ui.ws_tunnel.mode = None;
-    ui.ws_tunnel.enabled = false;
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    assert_eq!(config.adaptive.ws_tunnel_mode, WsTunnelMode::Off);
-}
-
-#[test]
-fn warp_rules_without_route_hosts_do_not_create_upstream_group() {
-    let mut ui = minimal_ui();
-    ui.warp.enabled = true;
-    ui.warp.route_mode = "rules".to_string();
-    ui.warp.route_hosts = String::new();
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(
-        config.groups.iter().all(|group| group.policy.label != "warp_routed"),
-        "ordinary empty WARP rules must not route all traffic",
-    );
-}
-
-#[test]
-fn amneziawg_route_marker_creates_all_traffic_upstream_group() {
-    let mut ui = minimal_ui();
-    ui.warp.enabled = true;
-    ui.warp.route_mode = "rules".to_string();
-    ui.warp.route_hosts = "__ripdpi_awg_all_traffic__".to_string();
-    ui.warp.local_socks_port = 10809;
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    let warp_group = config
-        .groups
-        .iter()
-        .find(|group| group.policy.label == "warp_routed")
-        .expect("AWG egress marker must create a WARP-routed upstream group");
-
-    assert!(warp_group.matches.filters.host_filters_empty());
-    assert_eq!(
-        warp_group.policy.ext_socks.map(|upstream| upstream.addr),
-        Some(std::net::SocketAddr::from(([127, 0, 0, 1], 10809))),
-    );
-}
-
-#[test]
-fn ui_http_strategy_enables_delay_connect() {
-    let mut ui = minimal_ui();
-    ui.protocols.desync_http = true;
-    ui.protocols.desync_udp = false;
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(config.network.delay_conn);
-}
-
-#[test]
-fn ui_host_filters_enable_delay_connect() {
-    let mut ui = minimal_ui();
-    ui.hosts.mode = "whitelist".to_string();
-    ui.hosts.entries = Some("example.com".to_string());
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(config.network.delay_conn);
-}
-
-#[test]
-fn ui_udp_only_strategy_keeps_delay_connect_disabled() {
-    let mut ui = minimal_ui();
-    ui.protocols.desync_http = false;
-    ui.protocols.desync_https = false;
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(!config.network.delay_conn);
-}
-
-#[test]
-fn udp_associate_unset_keeps_network_udp_enabled() {
-    let ui = minimal_ui();
-    assert_eq!(ui.protocols.udp_associate_enabled, None, "fixture must leave the switch unset");
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(config.network.udp, "unset udpAssociateEnabled must preserve the always-on default");
-}
-
-#[test]
-fn udp_associate_disabled_clears_network_udp() {
-    let mut ui = minimal_ui();
-    ui.protocols.udp_associate_enabled = Some(false);
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(!config.network.udp, "udpAssociateEnabled=false must disable SOCKS5 UDP ASSOCIATE");
-}
-
-#[test]
-fn udp_associate_enabled_sets_network_udp() {
-    let mut ui = minimal_ui();
-    ui.protocols.udp_associate_enabled = Some(true);
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert!(config.network.udp, "udpAssociateEnabled=true must enable SOCKS5 UDP ASSOCIATE");
-}
-
-#[test]
-fn actionable_ui_strategy_synthesizes_detect_connect_plain_fallback_group() {
-    let config = runtime_config_from_ui(minimal_ui()).expect("runtime config");
-
-    // minimal_ui() has desync_http=true, desync_https=true, desync_udp=true
-    // -> TCP group + UDP group + adaptive_direct + CONNECT-detect fallback
-    assert_eq!(config.groups.len(), 4);
-    assert_eq!(config.groups[0].matches.detect, 0);
-    assert_eq!(config.groups[0].matches.proto, IS_HTTP | IS_HTTPS);
-    assert_eq!(config.groups[1].matches.proto, IS_UDP);
-    let adaptive_direct = &config.groups[2];
-    assert_ne!(adaptive_direct.matches.detect, 0);
-    assert_eq!(adaptive_direct.matches.proto, IS_HTTP | IS_HTTPS);
-    let fallback = &config.groups[3];
-    assert_eq!(fallback.matches.detect, DETECT_CONNECT);
-    assert!(fallback.actions.tcp_chain.is_empty());
-    assert!(fallback.actions.udp_chain.is_empty());
-    assert_eq!(fallback.matches.proto, 0);
-}
-
-#[test]
-fn relay_upstream_covers_all_generated_fallback_groups() {
-    let mut ui = minimal_ui();
-    ui.upstream_relay.enabled = true;
-    ui.upstream_relay.kind = "vless_reality".to_string();
-    ui.upstream_relay.local_socks_port = 11_980;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-    let relay = std::net::SocketAddr::from(([127, 0, 0, 1], 11_980));
-
-    assert!(config.groups.len() >= 7, "fixture must include UDP, adaptive, and runtime-preset fallback groups");
-    for group in &config.groups {
-        assert_eq!(
-            group.policy.ext_socks.map(|upstream| upstream.addr),
-            Some(relay),
-            "group {} must not bypass the configured relay",
-            group.id,
-        );
-    }
-}
-
-#[test]
-fn relay_upstream_preserves_warp_routed_group() {
-    let mut ui = minimal_ui();
-    ui.upstream_relay.enabled = true;
-    ui.upstream_relay.kind = "vless_reality".to_string();
-    ui.upstream_relay.local_socks_port = 11_980;
-    ui.warp.enabled = true;
-    ui.warp.route_mode = "rules".to_string();
-    ui.warp.route_hosts = "__ripdpi_awg_all_traffic__".to_string();
-    ui.warp.local_socks_port = 10_809;
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-    let warp_group = config.groups.iter().find(|group| group.policy.label == "warp_routed").expect("WARP-routed group");
-
-    assert_eq!(
-        warp_group.policy.ext_socks.map(|upstream| upstream.addr),
-        Some(std::net::SocketAddr::from(([127, 0, 0, 1], 10_809))),
-    );
-}
-
-#[test]
-fn tcp_and_udp_desync_produces_separate_groups() {
-    let mut ui = minimal_ui();
-    ui.protocols.desync_http = true;
-    ui.protocols.desync_https = true;
-    ui.protocols.desync_udp = true;
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert_eq!(config.groups.len(), 4);
-    assert_eq!(config.groups[0].matches.proto & IS_UDP, 0, "TCP group must not have IS_UDP");
-    assert_eq!(config.groups[0].matches.proto, IS_HTTP | IS_HTTPS);
-    assert_eq!(config.groups[1].matches.proto, IS_UDP);
-    assert!(config.groups[1].actions.tcp_chain.is_empty(), "UDP group should not carry tcp_chain");
-    assert_ne!(config.groups[2].matches.detect, 0);
-    assert_eq!(config.groups[2].matches.proto, IS_HTTP | IS_HTTPS);
-    assert_eq!(config.groups[3].matches.detect, DETECT_CONNECT);
-}
-
-#[test]
-fn udp_only_strategy_produces_single_action_group() {
-    let mut ui = minimal_ui();
-    ui.protocols.desync_http = false;
-    ui.protocols.desync_https = false;
-    ui.protocols.desync_udp = true;
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    // TCP group (proto=0, no actions) + UDP group + CONNECT fallback
-    assert_eq!(config.groups.len(), 3);
-    assert_eq!(config.groups[0].matches.proto, 0);
-    assert_eq!(config.groups[1].matches.proto, IS_UDP);
-    assert_eq!(config.groups[2].matches.detect, DETECT_CONNECT);
-}
-
-#[test]
-fn tcp_only_strategy_produces_two_groups() {
-    let mut ui = minimal_ui();
-    ui.protocols.desync_http = true;
-    ui.protocols.desync_https = true;
-    ui.protocols.desync_udp = false;
-    ui.chains.udp_steps.clear();
-
-    let config = runtime_config_from_ui(ui).expect("runtime config");
-
-    assert_eq!(config.groups.len(), 3);
-    assert_eq!(config.groups[0].matches.proto, IS_HTTP | IS_HTTPS);
-    assert_ne!(config.groups[1].matches.detect, 0);
-    assert_eq!(config.groups[1].matches.proto, IS_HTTP | IS_HTTPS);
-    assert_eq!(config.groups[2].matches.detect, DETECT_CONNECT);
-}
-
-// --- Validation edge-case tests ---
-
-#[test]
-fn port_overflow_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.listen.port = 100_000;
-    let err = runtime_config_from_ui(ui).expect_err("port overflow");
-    assert!(err.to_string().contains("port"), "expected port error, got: {err}");
-}
-
-#[test]
-fn zero_port_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.listen.port = 0;
-    let err = runtime_config_from_ui(ui).expect_err("zero port");
-    assert!(err.to_string().contains("port"), "expected port error, got: {err}");
-}
-
-#[test]
-fn loopback_listen_without_auth_is_accepted() {
-    let mut ui = minimal_ui();
-    ui.listen.ip = "::1".to_string();
-    ui.listen.auth_token = None;
-
-    let config = runtime_config_from_ui(ui).expect("loopback listener should not require auth");
-
-    assert_eq!(config.network.listen.listen_ip.to_string(), "::1");
-    assert!(config.network.listen.auth_token.is_none());
-}
-
-#[test]
-fn non_loopback_listen_without_auth_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.listen.ip = "192.0.2.10".to_string();
-    ui.listen.auth_token = None;
-
-    let err = runtime_config_from_ui(ui).expect_err("non-loopback listener without auth");
-
-    assert!(err.to_string().contains("listen.authToken"), "expected listen.authToken error, got: {err}");
-}
-
-#[test]
-fn unspecified_listen_without_auth_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.listen.ip = "0.0.0.0".to_string();
-    ui.listen.auth_token = Some(String::new());
-
-    let err = runtime_config_from_ui(ui).expect_err("unspecified listener without auth");
-
-    assert!(err.to_string().contains("listen.authToken"), "expected listen.authToken error, got: {err}");
-}
-
-#[test]
-fn non_loopback_listen_with_auth_is_accepted() {
-    let mut ui = minimal_ui();
-    ui.listen.ip = "192.0.2.10".to_string();
-    ui.listen.auth_token = Some("local-lan-proxy-token".to_string());
-
-    let config = runtime_config_from_ui(ui).expect("non-loopback listener with auth");
-
-    assert_eq!(config.network.listen.listen_ip.to_string(), "192.0.2.10");
-    assert_eq!(config.network.listen.auth_token.as_deref(), Some("local-lan-proxy-token"));
-}
-
-#[test]
-fn zero_max_connections_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.listen.max_connections = 0;
-    let err = runtime_config_from_ui(ui).expect_err("zero max connections");
-    assert!(err.to_string().contains("maxConnections"), "expected maxConnections error, got: {err}");
-}
-
-#[test]
-fn negative_buffer_size_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.listen.buffer_size = -1;
-    let err = runtime_config_from_ui(ui).expect_err("negative buffer size");
-    assert!(err.to_string().contains("bufferSize"), "expected bufferSize error, got: {err}");
-}
-
-#[test]
-fn default_ttl_overflow_when_custom_ttl_enabled() {
-    let mut ui = minimal_ui();
-    ui.listen.custom_ttl = true;
-    ui.listen.default_ttl = 300;
-    let err = runtime_config_from_ui(ui).expect_err("ttl overflow");
-    assert!(err.to_string().contains("defaultTtl"), "expected defaultTtl error, got: {err}");
-}
-
-#[test]
-fn zero_default_ttl_when_custom_ttl_enabled() {
-    let mut ui = minimal_ui();
-    ui.listen.custom_ttl = true;
-    ui.listen.default_ttl = 0;
-    let err = runtime_config_from_ui(ui).expect_err("zero ttl");
-    assert!(err.to_string().contains("defaultTtl"), "expected defaultTtl error, got: {err}");
-}
-
-#[test]
-fn host_autolearn_enabled_without_store_path_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.host_autolearn.enabled = true;
-    ui.host_autolearn.store_path = None;
-    let err = runtime_config_from_ui(ui).expect_err("missing store path");
-    assert!(err.to_string().contains("storePath"), "expected storePath error, got: {err}");
-}
-
-// --- TCP chain validation tests ---
-
-#[test]
-fn tlsrec_after_send_step_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("split", "host+1"), tcp_step("tlsrec", "host+1")];
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("tlsrec after send");
-    assert!(err.to_string().contains("tlsrec"), "expected tlsrec error, got: {err}");
-}
-
-#[test]
-fn empty_tcp_chain_is_accepted() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![];
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("empty chain should be valid");
-    assert!(config.groups[0].actions.tcp_chain.is_empty());
-}
-
-#[test]
-fn tlsrec_before_send_step_is_accepted() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("tlsrec", "host+1"), tcp_step("disorder", "host+1")];
-    runtime_config_from_payload(ui_payload(ui)).expect("tlsrec before send should be valid");
-}
-
-#[test]
-fn fakeddisorder_not_last_is_rejected() {
-    let mut ui = minimal_ui();
-    ui.chains.tcp_steps = vec![tcp_step("fakeddisorder", "host+1"), tcp_step("split", "endhost")];
-    let err = runtime_config_from_payload(ui_payload(ui)).expect_err("non-terminal fakeddisorder");
-    assert!(err.to_string().contains("fakeddisorder"), "expected fakeddisorder error, got: {err}");
-}
-
-#[test]
-fn ui_config_maps_window_clamp() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.window_clamp = Some(2);
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("valid config");
-    assert_eq!(config.groups[0].actions.window_clamp, Some(2));
-}
-
-#[test]
-fn ui_config_maps_strip_timestamps() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.strip_timestamps = true;
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("valid config");
-    assert!(config.groups[0].actions.strip_timestamps);
-}
-
-#[test]
-fn ui_config_maps_quic_bind_low_port() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.quic_bind_low_port = true;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert!(config.groups[0].actions.quic_bind_low_port);
-}
-
-#[test]
-fn ui_config_maps_quic_migrate() {
-    let mut ui = minimal_ui();
-    ui.fake_packets.quic_migrate_after_handshake = true;
-
-    let config = runtime_config_from_payload(ui_payload(ui)).expect("runtime config");
-
-    assert!(config.groups[0].actions.quic_migrate_after_handshake);
-}
-
-#[test]
-fn ui_payload_maps_strategy_evolution_settings() {
-    let payload = serde_json::json!({
-        "kind": "ui",
-        "schemaVersion": 2,
-        "adaptiveFallback": {
-            "strategyEvolution": true,
-            "evolutionEpsilon": 0.2,
-            "evolutionExperimentTtlMs": 45_000,
-            "evolutionDecayHalfLifeMs": 1_800_000,
-            "evolutionCooldownAfterFailures": 5,
-            "evolutionCooldownMs": 120_000,
-            "futureAdaptiveField": { "version": 2 }
-        },
-        "futureTopLevel": true
-    });
-
-    let parsed = parse_proxy_config_json(&payload.to_string()).expect("additive config");
-    let config = runtime_config_envelope_from_payload(parsed).expect("runtime config").config;
-
-    assert!(config.adaptive.strategy_evolution);
-    assert_eq!(config.adaptive.evolution_epsilon_permil, 200);
-    assert_eq!(config.adaptive.evolution_experiment_ttl_ms, 45_000);
-    assert_eq!(config.adaptive.evolution_decay_half_life_ms, 1_800_000);
-    assert_eq!(config.adaptive.evolution_cooldown_after_failures, 5);
-    assert_eq!(config.adaptive.evolution_cooldown_ms, 120_000);
-}
-
-#[test]
-fn missing_strategy_evolution_settings_keep_current_defaults() {
-    let parsed = parse_proxy_config_json(r#"{"kind":"ui","schemaVersion":2,"adaptiveFallback":{}}"#)
-        .expect("current sparse config");
-    let config = runtime_config_envelope_from_payload(parsed).expect("runtime config").config;
-
-    assert!(!config.adaptive.strategy_evolution);
-    assert_eq!(config.adaptive.evolution_epsilon_permil, 100);
-    assert_eq!(config.adaptive.evolution_experiment_ttl_ms, 30_000);
-    assert_eq!(config.adaptive.evolution_decay_half_life_ms, 3_600_000);
-    assert_eq!(config.adaptive.evolution_cooldown_after_failures, 3);
-    assert_eq!(config.adaptive.evolution_cooldown_ms, 300_000);
-}
-
-#[test]
-fn schema_v2_without_destination_routing_keeps_inert_runtime_policy() {
-    let parsed = parse_proxy_config_json(r#"{"kind":"ui","schemaVersion":2,"adaptiveFallback":{}}"#)
-        .expect("old schema-v2 payload");
-    let config = runtime_config_from_payload(parsed).expect("runtime config");
-
-    assert!(config.destination_routing.rules.is_empty());
-    assert_eq!(config.destination_routing.default_action, ripdpi_config::DestinationRoutingAction::Tunneled);
-    assert!(config.destination_routing.canonical_digest.is_empty());
-}
-
-#[test]
-fn destination_routing_round_trips_and_maps_every_wire_variant() {
-    let json = with_destination_routing_digest(serde_json::json!({
-        "kind": "ui",
-        "schemaVersion": 2,
-        "destinationRouting": {
-            "rules": [
-                {
-                    "action": "direct",
-                    "network": "both",
-                    "domains": [
-                        {"kind": "exact", "value": "example.com"},
-                        {"kind": "suffix", "value": "example.net"},
-                        {"kind": "geosite", "value": "ru"}
-                    ],
-                    "ipRanges": [
-                        {"kind": "cidr", "value": "192.0.2.0/24"},
-                        {"kind": "geo_ip", "value": "ru"}
-                    ],
-                    "destinationPorts": [{"start": 443, "endInclusive": 8443}]
-                },
-                {
-                    "action": "block",
-                    "network": "tcp",
-                    "domains": [{"kind": "exact", "value": "blocked.example"}]
-                },
-                {
-                    "action": "tunneled",
-                    "network": "udp",
-                    "destinationPorts": [{"start": 53, "endInclusive": 53}]
-                }
-            ],
-            "defaultAction": "tunneled"
-        }
-    }));
-    assert_eq!(
-        json["destinationRouting"]["canonicalDigest"],
-        "5c6ebcf0aff1d5a4b0c1885acc4c7b8533718d1791066105336a03c9e09e7d51"
-    );
-
-    let parsed = parse_proxy_config_json(&json.to_string()).expect("destination routing payload");
-    let reparsed =
-        parse_proxy_config_json(&serde_json::to_string(&parsed).expect("serialize destination routing payload"))
-            .expect("reparse destination routing payload");
-    assert_eq!(parsed, reparsed);
-
-    let config = runtime_config_from_payload(parsed).expect("runtime config");
-    assert_eq!(config.destination_routing.rules.len(), 3);
-    assert_eq!(config.destination_routing.rules[0].action, ripdpi_config::DestinationRoutingAction::Direct);
-    assert_eq!(config.destination_routing.rules[0].network, ripdpi_config::DestinationRoutingNetwork::Both);
-    assert_eq!(config.destination_routing.rules[0].ip_ranges[1].kind, ripdpi_config::DestinationIpMatcherKind::GeoIp);
-    assert_eq!(config.destination_routing.rules[1].action, ripdpi_config::DestinationRoutingAction::Block);
-    assert_eq!(config.destination_routing.rules[2].network, ripdpi_config::DestinationRoutingNetwork::Udp);
-}
-
-#[test]
-fn destination_routing_rejects_unknown_enum_and_rule_fields() {
-    let unknown_action =
-        r#"{"kind":"ui","schemaVersion":2,"destinationRouting":{"rules":[],"defaultAction":"future"}}"#;
-    let unknown_rule_field = r#"{"kind":"ui","schemaVersion":2,"destinationRouting":{"rules":[{"action":"block","network":"tcp","domains":[{"kind":"exact","value":"example.com"}],"unexpected":true}],"canonicalDigest":"digest"}}"#;
-
-    assert!(parse_proxy_config_json(unknown_action).is_err());
-    assert!(parse_proxy_config_json(unknown_rule_field).is_err());
-}
-
-#[test]
-fn destination_routing_rejects_invalid_rule_atomically_during_conversion() {
-    let invalid_rule = r#"{"kind":"ui","schemaVersion":2,"destinationRouting":{"rules":[{"action":"direct","network":"both","destinationPorts":[{"start":8443,"endInclusive":443}]}],"defaultAction":"tunneled","canonicalDigest":"0000000000000000000000000000000000000000000000000000000000000000"}}"#;
-    let parsed = parse_proxy_config_json(invalid_rule).expect("structurally valid payload");
-
-    let error = runtime_config_from_payload(parsed).expect_err("invalid rule must reject the whole policy");
-    assert!(error.to_string().contains("invalid destination port range"));
-}
-
-#[test]
-fn destination_routing_rejects_direct_and_block_defaults_even_without_rules() {
-    for action in ["direct", "block"] {
-        let json = format!(
-            r#"{{"kind":"ui","schemaVersion":2,"destinationRouting":{{"rules":[],"defaultAction":"{action}"}}}}"#
-        );
-        let parsed = parse_proxy_config_json(&json).expect("structurally valid payload");
-
-        let error = runtime_config_from_payload(parsed).expect_err("unmatched destinations must always tunnel");
-        assert!(error.to_string().contains("defaultAction must be tunneled"));
-    }
-}
-
-#[test]
-fn destination_routing_empty_policy_rejects_nonempty_digest() {
-    let json = r#"{"kind":"ui","schemaVersion":2,"destinationRouting":{"rules":[],"defaultAction":"tunneled","canonicalDigest":"0000000000000000000000000000000000000000000000000000000000000000"}}"#;
-    let parsed = parse_proxy_config_json(json).expect("structurally valid payload");
-
-    let error = runtime_config_from_payload(parsed).expect_err("empty policy identity must be canonical");
-    assert!(error.to_string().contains("must be empty when rules are absent"));
-}
-
-#[test]
-fn destination_routing_rejects_noncanonical_matchers_and_stale_digest() {
-    let malformed_matchers = [
-        serde_json::json!({"domains": [{"kind": "exact", "value": "Example.com"}]}),
-        serde_json::json!({"domains": [{"kind": "suffix", "value": ".example.com"}]}),
-        serde_json::json!({"domains": [{"kind": "geosite", "value": "RU"}]}),
-        serde_json::json!({"ipRanges": [{"kind": "cidr", "value": "192.0.2.1/24"}]}),
-        serde_json::json!({"ipRanges": [{"kind": "geo_ip", "value": "RU"}]}),
-    ];
-    for matchers in malformed_matchers {
-        let mut rule = serde_json::json!({"action": "direct", "network": "both"});
-        rule.as_object_mut().expect("rule object").extend(matchers.as_object().expect("matcher object").clone());
-        let json = serde_json::json!({
-            "kind": "ui",
-            "schemaVersion": 2,
-            "destinationRouting": {
-                "rules": [rule],
-                "defaultAction": "tunneled",
-                "canonicalDigest": "0000000000000000000000000000000000000000000000000000000000000000"
-            }
-        });
-        let parsed = parse_proxy_config_json(&json.to_string()).expect("structurally valid payload");
-        assert!(runtime_config_from_payload(parsed).is_err());
-    }
-
-    let json = serde_json::json!({
-        "kind": "ui",
-        "schemaVersion": 2,
-        "destinationRouting": {
-            "rules": [{
-                "action": "direct",
-                "network": "tcp",
-                "domains": [{"kind": "exact", "value": "example.com"}]
-            }],
-            "defaultAction": "tunneled",
-            "canonicalDigest": "0000000000000000000000000000000000000000000000000000000000000000"
-        }
-    });
-    let parsed = parse_proxy_config_json(&json.to_string()).expect("structurally valid payload");
-    let error = runtime_config_from_payload(parsed).expect_err("forged digest must reject the whole policy");
-    assert!(error.to_string().contains("does not match"));
-
-    let valid = with_destination_routing_digest(json);
-    let uppercase_digest =
-        valid["destinationRouting"]["canonicalDigest"].as_str().expect("digest").to_ascii_uppercase();
-    let mut uppercase = valid;
-    uppercase["destinationRouting"]["canonicalDigest"] = serde_json::Value::String(uppercase_digest);
-    let parsed = parse_proxy_config_json(&uppercase.to_string()).expect("structurally valid payload");
-    let error = runtime_config_from_payload(parsed).expect_err("uppercase digest must reject the whole policy");
-    assert!(error.to_string().contains("lowercase hexadecimal"));
-}
-
-#[test]
-fn destination_routing_rejects_structural_bounds_atomically() {
-    let rules = (0..=256)
-        .map(|index| {
-            serde_json::json!({
-                "action": "direct",
-                "network": "tcp",
-                "domains": [{"kind": "exact", "value": format!("r{index}.example")}]
-            })
-        })
-        .collect::<Vec<_>>();
-    let json = serde_json::json!({
-        "kind": "ui",
-        "schemaVersion": 2,
-        "destinationRouting": {
-            "rules": rules,
-            "defaultAction": "tunneled",
-            "canonicalDigest": "0000000000000000000000000000000000000000000000000000000000000000"
-        }
-    });
-    let parsed = parse_proxy_config_json(&json.to_string()).expect("structurally valid payload");
-    let error = runtime_config_from_payload(parsed).expect_err("oversized policy must reject atomically");
-    assert!(error.to_string().contains("rules exceeds 256"));
-
-    let domains = (0..=256)
-        .map(|index| serde_json::json!({"kind": "exact", "value": format!("d{index}.example")}))
-        .collect::<Vec<_>>();
-    let json = serde_json::json!({
-        "kind": "ui",
-        "schemaVersion": 2,
-        "destinationRouting": {
-            "rules": [{"action": "direct", "network": "tcp", "domains": domains}],
-            "defaultAction": "tunneled",
-            "canonicalDigest": "0000000000000000000000000000000000000000000000000000000000000000"
-        }
-    });
-    let parsed = parse_proxy_config_json(&json.to_string()).expect("structurally valid payload");
-    let error = runtime_config_from_payload(parsed).expect_err("oversized field must reject atomically");
-    assert!(error.to_string().contains("domains exceeds 256"));
-
-    let json = serde_json::json!({
-        "kind": "ui",
-        "schemaVersion": 2,
-        "destinationRouting": {
-            "rules": [{
-                "action": "direct",
-                "network": "tcp",
-                "domains": [{"kind": "geosite", "value": "a".repeat(254)}]
-            }],
-            "defaultAction": "tunneled",
-            "canonicalDigest": "0000000000000000000000000000000000000000000000000000000000000000"
-        }
-    });
-    let parsed = parse_proxy_config_json(&json.to_string()).expect("structurally valid payload");
-    let error = runtime_config_from_payload(parsed).expect_err("oversized token must reject atomically");
-    assert!(error.to_string().contains("non-canonical domain matcher"));
-}
-
-#[test]
-fn destination_routing_accepts_exact_aggregate_and_canonical_bounds() {
-    let policy = exact_aggregate_boundary_policy();
-
-    runtime_config_from_destination_policy(policy)
-        .expect("exact rule, matcher, and canonical-size bounds must be accepted");
-}
-
-#[test]
-fn destination_routing_rejects_aggregate_and_canonical_overflow() {
-    let mut aggregate = exact_aggregate_boundary_policy();
-    aggregate.rules[0].domains.push(ProxyUiDestinationDomainMatcher {
-        kind: ProxyUiDestinationDomainMatcherKind::Geosite,
-        value: format!("{:a<63}", "aggregate-overflow"),
-    });
-    aggregate.canonical_digest = "0".repeat(64);
-    let error = runtime_config_from_destination_policy(aggregate).expect_err("1,025 matchers must reject atomically");
-    assert!(error.to_string().contains("matcher count exceeds 1024"));
-
-    let mut canonical = exact_aggregate_boundary_policy();
-    canonical.rules[0].domains[0].value.push('a');
-    canonical.canonical_digest = "0".repeat(64);
-    let error = runtime_config_from_destination_policy(canonical)
-        .expect_err("65,537 canonical characters must reject atomically");
-    assert!(error.to_string().contains("exceeds 65536"));
-}
-
-#[test]
-fn destination_routing_field_limits_cover_ip_ranges_and_ports() {
-    let exact_rule = ProxyUiDestinationRoutingRule {
-        action: ProxyUiDestinationRoutingAction::Direct,
-        network: ProxyUiDestinationRoutingNetwork::Both,
-        domains: (0..256)
-            .map(|index| ProxyUiDestinationDomainMatcher {
-                kind: ProxyUiDestinationDomainMatcherKind::Geosite,
-                value: format!("d{index}"),
-            })
-            .collect(),
-        ip_ranges: (0..256)
-            .map(|index| ProxyUiDestinationIpMatcher {
-                kind: ProxyUiDestinationIpMatcherKind::Cidr,
-                value: format!("10.0.{index}.0/24"),
-            })
-            .collect(),
-        destination_ports: (1..=256)
-            .map(|port| ProxyUiDestinationPortRange { start: port, end_inclusive: port })
-            .collect(),
-    };
-    runtime_config_from_destination_policy(destination_policy_with_digest(vec![exact_rule.clone()]))
-        .expect("256 entries in every matcher field must be accepted");
-
-    let mut ip_overflow = exact_rule.clone();
-    ip_overflow.domains.clear();
-    ip_overflow.destination_ports.clear();
-    ip_overflow.ip_ranges.push(ProxyUiDestinationIpMatcher {
-        kind: ProxyUiDestinationIpMatcherKind::Cidr,
-        value: "10.1.0.0/24".to_string(),
-    });
-    let error = runtime_config_from_destination_policy(destination_policy_with_digest(vec![ip_overflow]))
-        .expect_err("257 IP ranges must reject atomically");
-    assert!(error.to_string().contains("ipRanges exceeds 256"));
-
-    let mut port_overflow = exact_rule;
-    port_overflow.domains.clear();
-    port_overflow.ip_ranges.clear();
-    port_overflow.destination_ports.push(ProxyUiDestinationPortRange { start: 257, end_inclusive: 257 });
-    let error = runtime_config_from_destination_policy(destination_policy_with_digest(vec![port_overflow]))
-        .expect_err("257 port ranges must reject atomically");
-    assert!(error.to_string().contains("destinationPorts exceeds 256"));
-}
+#[path = "tests_support.rs"]
+mod tests_support;
+
+use tests_support::*;
+
+#[path = "tests_conversion.rs"]
+mod tests_conversion;
+#[path = "tests_runtime.rs"]
+mod tests_runtime;
